requirements.txt
altair==5.0.1
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
easyocr==1.7.0
filelock==3.12.0
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.15.1
idna==3.4
imageio==2.31.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
networkx==2.6.3
ninja==1.11.1
numpy==1.21.6
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
pip==20.1.1
pkgutil-resolve-name==1.3.10
protobuf==4.23.2
pyarrow==12.0.0
pyclipper==1.3.0.post4
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
setuptools==47.1.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
streamlit==1.23.1
tenacity==8.2.2
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.2
tqdm==4.65.0
transformers==4.29.2
typing-extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
wget==3.2
zipp==3.15.0
